fix: unify 'out of sync' logic to composable
