+++ frontend/app/src/components/profitloss/ReportGenerator.vue
@@ -2,13 +2,8 @@
 import type { ProfitLossReportPeriod } from '@/types/reports';
 import RangeSelector from '@/components/helper/date/RangeSelector.vue';
 import CardTitle from '@/components/typography/CardTitle.vue';
-import { useRefWithDebounce } from '@/composables/ref';
-import {
-  useHistoryQueryIndicatorSettings,
-} from '@/modules/dashboard/history-events/composables/use-history-query-indicator-settings';
-import { useHistoryEventsStatus } from '@/modules/history/events/use-history-events-status';
+import { useTransactionStatusCheck } from '@/modules/dashboard/history-events/composables/use-transaction-status-check';
 import { Routes } from '@/router/routes';
-import { useHistoryStore } from '@/store/history';
 
 const emit = defineEmits<{
   (e: 'generate', data: ProfitLossReportPeriod): void;
@@ -17,52 +12,11 @@ const emit = defineEmits<{
 }>();
 
 const { t } = useI18n({ useScope: 'global' });
-const router = useRouter();
 
-const historyStore = useHistoryStore();
-const { transactionStatusSummary } = storeToRefs(historyStore);
-const { processing: rawProcessing } = useHistoryEventsStatus();
-const processing = useRefWithDebounce(rawProcessing, 400);
-const { minOutOfSyncPeriodMs } = useHistoryQueryIndicatorSettings();
+const { isOutOfSync, navigateToHistory, processing } = useTransactionStatusCheck();
 
 const range = ref<ProfitLossReportPeriod>({ end: 0, start: 0 });
 const valid = ref<boolean>(false);
-
-const isOutOfSync = computed<boolean>(() => {
-  const status = get(transactionStatusSummary);
-  if (!isDefined(status)) {
-    return false;
-  }
-
-  const { evmLastQueriedTs = 0, exchangesLastQueriedTs = 0, hasEvmAccounts = false, hasExchangesAccounts = false } = status;
-
-  // Only check if user has accounts
-  if (!hasEvmAccounts && !hasExchangesAccounts) {
-    return false;
-  }
-
-  const now = Date.now();
-  const minOutOfSyncMs = get(minOutOfSyncPeriodMs);
-
-  // Check EVM if user has EVM accounts
-  if (hasEvmAccounts && evmLastQueriedTs > 0) {
-    const evmLastQueriedMs = evmLastQueriedTs * 1000;
-    if (now - evmLastQueriedMs >= minOutOfSyncMs) {
-      return true;
-    }
-  }
-
-  // Check exchanges if user has exchange accounts
-  if (hasExchangesAccounts && exchangesLastQueriedTs > 0) {
-    const exchangesLastQueriedMs = exchangesLastQueriedTs * 1000;
-    if (now - exchangesLastQueriedMs >= minOutOfSyncMs) {
-      return true;
-    }
-  }
-
-  // Also consider it out of sync if never queried
-  return (hasEvmAccounts && evmLastQueriedTs === 0) || (hasExchangesAccounts && exchangesLastQueriedTs === 0);
-});
 
 const canGenerate = computed<boolean>(() => get(valid) && !get(processing) && !get(isOutOfSync));
 
@@ -76,10 +30,6 @@ function exportReportData(): void {
 
 function importReportData(): void {
   emit('import-data');
-}
-
-function navigateToHistory(): void {
-  router.push('/history');
 }
 
 const accountSettingsRoute = Routes.SETTINGS_ACCOUNTING;
